Python_Model/pythonProject/requirements.txt
flask
flask-cors
pillow
numpy
matplotlib
opencv-python
tensorflow
tf-keras-vis
cloudinary
dotenv